refactor: 프로필 관련 query를 커스텀 훅으로 분리
